drivers: mmc: sdhci-msm: Disable clock scaling for Sony devices
-- SONY SPECIFIC --

This is actually an hack more than something clean. Explaining it,
Sony devices actually need it because Samsung MAG2GC will act bad
with clock scaling: it's unable to make a good clock decision based
on load activities.

The only way to disable clock scaling for this particular eMMC is,
unfortunately, to disable it controller-wide, because there are
no MMC-specific quirks to use for this.

Signed-off-by: mostafa-z <mostafazarghami@gmail.com>