config enable DEBUG_ALIGN_RODATA
Align linker sections:
If this option is enabled, sections that may potentially be marked as read only or non-executable will be
aligned up to the section size of the kernel. This prevents sections from being split into pages and avoids a
potential TLB penalty. The downside is an increase in alignment and potentially wasted space. Turn on this option if
performance is more important than memory pressure.

Signed-off-by: mostafa-z <mostafazarghami@gmail.com>